% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in core.R
\name{init_monitor}
\alias{init_monitor}
\title{Initialize the environment monitor}
\usage{
init_monitor(path)
}

\arguments{
\item{path}{Path to the JSON file where environment data should be written}
}

\description{
Initialize the environment monitor
}

